+++ src/argon2/_api.py
@@ -5,7 +5,7 @@ from __future__ import absolute_import, division, print_function
 import os
 
 from ._util import (
-    check_types, error_to_str, ffi, lib, get_encoded_len, Type, NoneType,
+    _error_to_str, ffi, lib, _get_encoded_len, Type
 )
 from .exceptions import VerificationError, HashingError
 
@@ -44,11 +44,11 @@ def hash_password(password, salt=None,
     :param bytes salt: A salt_.  Should be random and different for each
         password.  Will generate a random salt for you if left ``None``
         (recommended).
-    :param int time_cost: Defines the amount of computation realized and
-        therefore the execution time, given in number of iterations.
-    :param int memory_cost: Defines the memory usage, given in kibibytes_.
-    :param int parallelism: Defines the number of parallel threads (*changes*
-        the resulting hash value).
+    :param int time_cost: Amount of computation realized and therefore the
+        execution time, given in number of iterations.
+    :param int memory_cost: Memory usage, given in kibibytes_.
+    :param int parallelism: Number of parallel threads (*changes* the resulting
+        hash value).
     :param int hash_len: Length of the hash in bytes.
     :param Type type: Which Argon2 variant to use.  In doubt use the default
         :attr:`Type.I` which is better suited for passwords.
@@ -79,25 +79,13 @@ def hash_password_raw(password, salt=None,
 
 def _hash(password, salt, time_cost, memory_cost, parallelism, hash_len, type,
           encoded):
-    e = check_types(
-        password=(password, bytes),
-        salt=(salt, (bytes, NoneType)),
-        time_cost=(time_cost, int),
-        memory_cost=(memory_cost, int),
-        parallelism=(parallelism, int),
-        hash_len=(hash_len, int),
-        type=(type, Type),
-        encoded=(encoded, bool),
-    )
-    if e:
-        raise TypeError(e)
     if salt is None:
         salt = os.urandom(DEFAULT_RANDOM_SALT_LENGTH)
 
     raw_buf = encoded_buf = ffi.NULL
     raw_len = encoded_len = 0
     if encoded:
-        encoded_len = get_encoded_len(hash_len, len(salt))
+        encoded_len = _get_encoded_len(hash_len, len(salt))
         encoded_buf = ffi.new("char[]", encoded_len)
     else:
         raw_len = hash_len
@@ -112,7 +100,7 @@ def _hash(password, salt, time_cost, memory_cost, parallelism, hash_len, type,
         type.value,
     )
     if rv != lib.ARGON2_OK:
-        raise HashingError(error_to_str(rv))
+        raise HashingError(_error_to_str(rv))
 
     return (
         ffi.string(encoded_buf) if encoded_len != 0
@@ -130,17 +118,10 @@ def verify_password(hash, password, type=Type.I):
         in *hash*.
     :param Type type: Type for *hash*.
 
-    :return: ``True`` on success, throw exception otherwise.
+    :return: ``True`` on success, raise
+            :exc:`~argon2.exceptions.VerificationError` otherwise.
     :rtype: bool
     """
-    e = check_types(
-        password=(password, bytes),
-        hash=(hash, bytes),
-        type=(type, Type),
-    )
-    if e:
-        raise TypeError(e)
-
     rv = lib.argon2_verify(
         ffi.new("char[]", hash),
         ffi.new("char[]", password),
@@ -150,4 +131,4 @@ def verify_password(hash, password, type=Type.I):
     if rv == lib.ARGON2_OK:
         return True
     else:
-        raise VerificationError(error_to_str(rv))
+        raise VerificationError(_error_to_str(rv))